requirements.txt
requests==2.31.0
python-telegram-bot==20.7
youtube-transcript-api==0.6.2
pytube==15.0.0
python-dotenv==1.0.0